+++ src/main/java/com/jakewharton/trakt/entities/TvShow.java
@@ -2,23 +2,25 @@ package com.jakewharton.trakt.entities;
 
 import java.util.Date;
 import java.util.List;
+import com.google.gson.annotations.SerializedName;
 import com.jakewharton.trakt.TraktEntity;
 
 public final class TvShow extends MediaBase implements TraktEntity {
 	private static final long serialVersionUID = 862473930551420996L;
 
-	private Date firstAired;
+	@SerializedName("first_aired") private Date firstAired;
 	private String country;
 	private String overview;
 	private Integer runtime;
 	private String network;
-	private String airDay; //TODO: enum
-	private String airTime;
+	@SerializedName("air_day") private String airDay; //TODO: enum
+	@SerializedName("air_time") private String airTime;
 	private String certification; //TODO: enum
-	private String tvdbId;
-	private String tvrageId;
+	@SerializedName("tvdb_id") private String tvdbId; //TODO: convert to Integer
+	@SerializedName("tvrage_id") private String tvrageId; //TODO: convert to Integer
 	private List<TvShowEpisode> episodes;
-	private List<TvShowEpisode> topEpisodes;
+	@SerializedName("top_episodes") private List<TvShowEpisode> topEpisodes;
+	private List<TvShowSeason> seasons;
 	
 	public Date getFirstAired() {
 		return this.firstAired;
@@ -55,5 +57,8 @@ public final class TvShow extends MediaBase implements TraktEntity {
 	}
 	public List<TvShowEpisode> getTopEpisodes() {
 		return this.topEpisodes;
+	}
+	public List<TvShowSeason> getSeasons() {
+		return this.seasons;
 	}
 }