+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Change Log
 ==========
 
+1.0.1 *(In Development)*
+------------------------
+
+ * Fix deserialization of fields whose names contained underscores.
+ * `getEpisodes()` in `TvShowSeason` is now a complex type that can hold both
+   a count and an episode number list.
+
+
 1.0.0 *(2011-07-01)*
 --------------------